css/css-fonts/parsing/font-variant-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-variant with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-prop">
<meta name="assert" content="font-variant supports only the specified grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("font-variant", "auto");
test_invalid_value("font-variant", "none normal");
test_invalid_value("font-variant", "sub super");
</script>
</body>
</html>
